DefaultDataTemplateSelctor3D looks for resources with base types as keys to mirror DataTemplate's behavior. Fixed spellings.
