+++ Source/Examples/WPF/ExampleBrowser/ExampleBrowser.csproj
@@ -98,9 +98,10 @@
     <Compile Include="Examples\DataTemplate\BindingConverter.cs" />
     <Compile Include="Examples\DataTemplate\BindingExtensions.cs" />
     <Compile Include="Examples\DataTemplate\DataTemplateSelector3D.cs" />
-    <Compile Include="Examples\DataTemplate\DefaultDatTemplateSelctor3D.cs" />
+    <Compile Include="Examples\DataTemplate\DefaultDataTemplateSelctor3D.cs" />
     <Compile Include="Examples\DataTemplate\CubeElement.cs" />
     <Compile Include="Examples\DataTemplate\Extensions.cs" />
+    <Compile Include="Examples\DataTemplate\ModelElement2.cs" />
     <Compile Include="Examples\DataTemplate\ModelElement1.cs" />
     <Compile Include="Examples\DataTemplate\ModelElement.cs" />
     <Compile Include="Examples\DataTemplate\MultiBindingConverter.cs" />
