+++ Source/Examples/WPF/ExampleBrowser/Examples/DataTemplate/ModelElement2.cs
@@ -1,0 +1,6 @@
+﻿namespace DataTemplateDemo
+{
+    public class ModelElement2 : ModelElement
+    {
+    }
+}
